ICU-23272 Fix long skeleton crash while parsing

+++ icu4c/source/i18n/unicode/numberformatter.h
@@ -1232,7 +1232,7 @@ class U_I18N_API StringProp : public UMemory {
     StringProp &operator=(StringProp &&src) noexcept;
 
     /** @internal */
-    int16_t length() const {
+    int32_t length() const {
         return fLength;
     }
 
@@ -1250,7 +1250,7 @@ class U_I18N_API StringProp : public UMemory {
 
   private:
     char *fValue;
-    int16_t fLength;
+    int32_t fLength;
     UErrorCode fError;
 
     StringProp() : fValue(nullptr), fLength(0), fError(U_ZERO_ERROR) {
